% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-agingdata.R
\docType{data}
\name{agingdata}
\alias{agingdata}
\title{Montpetit and Bergeman Longitudinal Study on Aging Data}
\format{
A data frame with 521 observations on the following 7 variables.
}
\usage{
agingdata
}
\description{
The data consist of 521 cases and 7 variables with 280 of the cases being complete. The 
variables are Education, Income, Perceived satisfaction of social support, Social coping, Total 
life events scale, Depression scale, and Self-rated help.
}
\references{
Montpetit A, Bergeman CS (2007). ``Dimensions of control: Mediational analyses of the stress Chealth relationship.'' \emph{Personality and Individual Differences,} {43}, 2237 - 2248, \doi{10.1016/j.paid.2007.07.003}.
}
\keyword{dataset}
